requirements.txt
# Dependencies for Render Service Keep-Alive & Monitoring
requests>=2.28.1
python-dotenv>=0.21.0
loguru>=0.6.0

# FastAPI dependencies for Vercel deployment
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
mangum>=0.17.0